requirements.txt
anyio==4.3.0
asgiref==3.8.1
click==8.1.7
fastapi==0.78.0
freezegun==1.5.1
h11==0.14.0
idna==3.7
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pydantic==1.10.15
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.19.1
typing_extensions==4.11.0
uvicorn==0.17.6
